refactor: unified per-object tool generator architecture
- Create PerObjectToolGeneratorService for shared object iteration logic
- Extract directRecordToolsFactory to record-crud/tool-factory/
- Consolidate 4 step configurator files into single workflowStepToolsFactory
- Simplify ToolService to use new generator (~220 lines removed)
- Simplify WorkflowToolWorkspaceService to use new generator (~150 lines removed)
- Delete redundant record-step-configurators/ directory

Total reduction: ~350+ lines of duplicated code
