+++ packages/twenty-server/src/modules/workflow/workflow-tools/workflow-tools.module.ts
@@ -1,8 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { RecordPositionModule } from 'src/engine/core-modules/record-position/record-position.module';
-import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
-import { WorkspaceCacheModule } from 'src/engine/workspace-cache/workspace-cache.module';
+import { ToolGeneratorModule } from 'src/engine/core-modules/tool-generator/tool-generator.module';
 import { WorkflowSchemaModule } from 'src/modules/workflow/workflow-builder/workflow-schema/workflow-schema.module';
 import { WorkflowVersionEdgeModule } from 'src/modules/workflow/workflow-builder/workflow-version-edge/workflow-version-edge.module';
 import { WorkflowVersionStepModule } from 'src/modules/workflow/workflow-builder/workflow-version-step/workflow-version-step.module';
@@ -19,8 +18,7 @@ import { WorkflowToolWorkspaceService } from './services/workflow-tool.workspace
     WorkflowTriggerModule,
     WorkflowSchemaModule,
     RecordPositionModule,
-    WorkspaceManyOrAllFlatEntityMapsCacheModule,
-    WorkspaceCacheModule,
+    ToolGeneratorModule,
   ],
   providers: [WorkflowToolWorkspaceService],
   exports: [WorkflowToolWorkspaceService],
